#2 1. Implementation of discount logic 2. Adding controller

+++ src/main/java/com/test/billcalculator/model/User.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class User {
     private UserType userType;
-    private int userTenure;
+    private float userTenure;
 }
